Improve Go compiler version detection
This commit adds new logic for detecting the Go compiler version used
when compiling the binary. It uses the schedinit function in the runtime
package to find the string identifier. This functionality works for
binaries compiled with 1.7 and newer.

PE, ELF and Mach-O binaries are supported. This will be used for x86
binaries only. The old logic is used as a fallback if this method fails
or not supported.