Fix pclntab search for Go 1.16
In Go release 1.16, some internal structure has changed. As part of this
change, the magic bytes used at the beginning of the table was also
changed. For PE binaries, gore uses these bytes to find the start of the
table. Because of this, gore could not find the table in PE binaries
compiled with 1.16.0 or newer. ELF and Mach-O are not affected.

Since the standard library provides logic for parsing the table, a
minimum version 1.16 is now required. Since logic to handle both types
of tables are not supported in older versions. Due to this, only 1.16
and newer will be supported.