Type parser re-write
This big commit has a new written parser for Go type. The logic has been
implemented for binaries compiled with Go 1.7 and later. The logic was
not backported to earlier versions due to the big difference in data
structures and the versions are so old that the old logic should be
fine.

As part of this, the code that handles 1.7 and newer was removed from
the old parser.