Handle files when external linker was used.
This adds logic that gets the runtime.text value from the moduledata
structure, and uses it to parse the pclntab. This ensures that the
function address are correct since the runtime.text is used as the base
address by the runtime.

Additionally, ELF files compiled as PIE files, does not have a specific
section for the pclntab. So a fallback logic has been added to search
for the table that is similar to what is used for PE files.